4.7.3 fix (test): TestCafe bisect script:
- `test/test_scripts/mapjs_bisect_testcafe.sh`:
  - Check whether replay script chosen.
  - Show feedback messages in console.
  - Add test mode so git commands and install can be disabled. This stops changes being reset, and also speeds up testing.
  - Fix boolean logic checks.

+++ test/test_scripts/mapjs_bisect_testcafe.sh
@@ -10,10 +10,14 @@
 # ------------
 
 # QUESTION can I move these two commands outside of this script into calling function?
-export PATH_REPLAY_SCRIPT=$DIR_PROJECTS/argmap/test/devtools-recordings/add_idea.json
+# export PATH_REPLAY_SCRIPT=$DIR_PROJECTS/argmap/test/devtools-recordings/add_idea.json
 
-source /home/s6mike/scripts/bash_aliases.sh                      # Functions to ensure clean build.
-source /home/s6mike/scripts/argmap_scripts/bash_aliases_mapjs.sh # Function to test rendering.
+# source /home/s6mike/scripts/bash_aliases.sh                      # Functions to ensure clean build.
+# source /home/s6mike/scripts/argmap_scripts/bash_aliases_mapjs.sh # Function to test rendering.
+
+TEST_MODE=false # false allows git commands to remove any changes.
+
+BISECT_START_TIMESTAMP=${BISECT_START_TIMESTAMP:-$(date "+%Y.%m.%d-%H.%M.%S")}
 
 # Add Hotfixes
 # ------------
@@ -30,40 +34,35 @@ source /home/s6mike/scripts/argmap_scripts/bash_aliases_mapjs.sh # Function to t
 # map-model.js _ fix
 # toolbar fix
 
-git stash apply || (
-  echo "'git stash apply' failed, exit 255 (abort bisect)"
-  exit 255
-)
+if [ "$TEST_MODE" = false ]; then # Only runs if not in test mode
 
-echo "'git stash apply' successful."
+  if git stash apply; then
+    echo "'git stash apply' successful."
+  else
+    echo "'git stash apply' failed, exit 255 (abort bisect)"
+    exit 255
+  fi
+
+fi
 
 # Build
 #------
 
-# npm install
-# wait
-# npm exec webpack
-# wait
-
 # webpack_build
-npm install --prefix "$PATH_MJS_HOME" --legacy-peer-deps
+if [ "$TEST_MODE" = false ]; then # Only runs if not in test mode
+  npm install --prefix "$PATH_MJS_HOME" --legacy-peer-deps
+fi
 
 # webpack_server_start # Dev server should already be started
-# wait
-
-# "testc": "npm run test-cafe",
-# "test-start": "npm run start && npm run test-cafe",
-# "stop": "killall node",
-# "start": "npm run pack && webpack-dev-server &",
-# "test-cafe": "testcafe 'chrome:headless --no-default-browser-check' $DIR_PROJECTS/argmap/test/devtools-recordings/Add_supporting_group.json",
-# "pack": "webpack",
-#  "server": "webpack-dev-server"
-
 # git add -u
 
 # TODO: Ideally want to read node_count_before and pass it into the JSON script as a variable to compare it to.
 #   See https://workflowy.com/#/33d0bdfaf875
 # node_count_before = document.getElementsByClassName('mapjs-node').length
+
+# Included in test (expecting 14):
+# node_count_after = document.getElementsByClassName('mapjs-node').length
+# exit if node_count_after > node_count_before then 0 else 1
 
 # Run Test
 # --------
@@ -78,32 +77,35 @@ npm install --prefix "$PATH_MJS_HOME" --legacy-peer-deps
 # else
 
 # Pre-test: Do any elements render?
+__test_mapjs_renders
+rendered=$?
 
-# npm exec testcafe 'chrome:headless --no-default-browser-check' $DIR_PROJECTS/argmap/test/devtools-recordings/map_renders.json >>"../git-bisect-logs/$BISECT_START_TIMESTAMP-testcafe-output.txt"
-# renders=$?
-
-# if [ $renders != 0 ]; then
-
-__test_mapjs_renders ""
-render_status=$?
-if [ $render_status == 0 ]; then
-  npm exec testcafe 'chrome:headless --no-default-browser-check' "$PATH_REPLAY_SCRIPT" >>"../git-bisect-logs/$BISECT_START_TIMESTAMP-testcafe-output.txt"
-  exit_status=$?
-else # if headless chrome fails to render any map nodes
+if [ $rendered = 0 ]; then # headless chrome fails to render any map nodes
+  # Run main test
+  echo "Render test succeeded, running testcafe script: $PATH_REPLAY_SCRIPT"
+  if [ "$PATH_REPLAY_SCRIPT" == "" ]; then
+    echo "No TestCafe PATH_REPLAY_SCRIPT set, setting exit status to 255 (bisect abort)"
+    exit_status=255
+    exit 255 # exits without git reset
+  else
+    npm exec testcafe 'chrome:headless --no-default-browser-check' "$PATH_REPLAY_SCRIPT" >>"../git-bisect-logs/$BISECT_START_TIMESTAMP-testcafe-output.txt"
+    exit_status=$?
+  fi
+else # render succeeds
   exit_status=125
-  echo "Map not rendered, setting exit status to 125 (bisect skip)."
+  echo "Rendering test failed, setting exit status to 125 (bisect skip)."
 fi
-
-# Included in test (expecting 14):
-# node_count_after = document.getElementsByClassName('mapjs-node').length
-# exit if node_count_after > node_count_before then 0 else 1
 
 # Bisect Teardown
 # ---------------
 
 # node_stop
 # This is necessary to allow checkout to work for next bisect step:
-git reset --hard
+
+# Only resets if not in test mode
+if [ "$TEST_MODE" = false ]; then # Won't in test mode
+  git reset --hard
+fi
 
 # Return test result
 echo "bisect exit status: $exit_status"
